refactor: triage and clean up unused type exports (#824)
Resolve all 8 unused type exports flagged by knip:
- Keep and annotate intentional public type exports (,
  , , ,
  ) using inline comment tags.
- Delete duplicate unused  in .
- Delete legacy unused  alias in .
- Remove internal barrel re-exports in .
- Refactor , , and
   to import RAG types directly from .
