+++ src/services/rag-types.ts
@@ -1,6 +1,7 @@
 /**
  * Represents a file that has been indexed in the File Search Store
  */
+// knip:keep — Intentional public API structurally consumed by RagIndexCache.files
 export interface IndexedFileEntry {
 	resourceName: string; // Gemini file resource name
 	contentHash: string; // SHA-256 hash for reliable change detection
